Rename Overview/index.jest.tsx to OverviewRoute.jest.tsx and refactor OverviewRoute tests to use new pattern
